# Add indexes to speed up views