### Jira link

[HMRC-811](https://transformuk.atlassian.net/browse/HMRC-811)

### What?

I have updated:

- Index on quota_definitions_oplog
- Index on quota_balance_events_oplog

In both cases an oid in descending order has been to an existing index which is mainly used to create the related view.

### Why?

I am doing this because:

Speeds up queries on 3 views (old and new time to count all rows):
- quota_definitions (9 seconds -> 0.35 seconds)
- bad_quota_associations (10 seconds -> 1.5 seconds)
- quota_balance_events (300 seconds -> 12 seconds)

In particular this should help to speed up quota search

### Have you? (optional)

- [ ] Added documentation for new apis
- [ ] Added new environment variables with correct values to all apps in all spaces

### Deployment risks (optional)

- Includes destructive/migratory actions which may effect critical data
- Changes an api that is used in production